Add files via upload
OFDM transmitter and receiver are implemented in GNU Radio on two different computer systems.
GRC files for transmitter and receiver are availabel with name Tx_Ofdm.grc and Rx_Ofdm.grc.
Generated OFDM signal is transmitted and recieved using USRP and data is recovered at the receiver.
Visualized the power spectrum of received signal and constellation diagram at the reciever. Output is added as .png file.